remove color from profile router
